moved sinon.format() to core
* commonjs-ified sinon.format
