+++ lib/sinon/assert.js
@@ -9,7 +9,6 @@
 "use strict";
 
 require("./match");
-require("./format");
 var sinon = require("./util/core");
 
 var slice = Array.prototype.slice;
