+++ lib/sinon/call.js
@@ -11,7 +11,6 @@
 "use strict";
 
 require("./match");
-require("./format");
 var sinon = require("./util/core");
 var slice = Array.prototype.slice;
