+++ lib/sinon/mock.js
@@ -13,7 +13,6 @@ require("./extend");
 require("./match");
 require("./spy");
 require("./stub");
-require("./format");
 var sinon = require("./util/core");
 
 var push = [].push;
