+++ lib/sinon/spy.js
@@ -10,7 +10,6 @@
 
 require("./extend");
 require("./call");
-require("./format");
 var sinon = require("./util/core");
 
 var push = Array.prototype.push;
