+++ lib/sinon/util/core/index.js
@@ -14,6 +14,8 @@ exports.create = require("./create");
 
 exports.deepEqual = require("./deep-equal");
 
+exports.format = require("./format");
+
 exports.functionName = require("./function-name");
 
 exports.functionToString = require("./function-to-string");
