+++ lib/sinon/util/fake_server.js
@@ -13,7 +13,6 @@
 
 require("./fake_xdomain_request");
 require("./fake_xml_http_request");
-require("../format");
 require("../log_error");
 
 var push = [].push;
